Move agent grant and workflow principal helpers into services
Service-specific support helpers now live with their owning services.

Canonical agent grant API:

    import "github.com/valon-technologies/gestalt/server/services/agents/agentgrant"

    grants, err := agentgrant.NewManager(secret)
    token, err := grants.Mint(agentgrant.Grant{...})

Canonical workflow principal API:

    import "github.com/valon-technologies/gestalt/server/services/workflows/workflowprincipal"

    principal := workflowprincipal.FromExecutionReference(ref)

The package names stay agentgrant and workflowprincipal; this commit moves the directories and rewrites imports without changing behavior.
